ref: Move event filters into libgeneral
