+++ general/Cargo.toml
@@ -12,7 +12,7 @@ debugid = { version = "0.5.1", features = ["serde"] }
 dynfmt = { version = "0.1.1", features = ["python", "curly"] }
 failure = "0.1.5"
 hmac = "0.7.0"
-insta = "0.7.4"
+insta = "0.9.0"
 itertools = "0.8.0"
 lazy_static = "1.2.0"
 maxminddb = "0.12.0"
